% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_Bag}
\alias{plot_Bag}
\title{Plot a GP_Bag model}
\usage{
plot_Bag(mod, x.test, y.test)
}
\arguments{
\item{mod}{A GP_Bag object.}

\item{x.test}{Data.frame of test set.}

\item{y.test}{Data.frame of test observations.}
}
\value{
ggplot. Returns a ggplot with mean prediction and error bounds
}
\description{
his function returns ggplot for a GP_Bag type object, thereby plotting a bagged
Gaussian Process regression model with mean prediction and error bounds, along
with the true function in green.
}
\details{
The \code{plot.GP_Bag} conveniently plots a \code{GP_Bag} object, so you
done't have to yourself.
}
\note{
mod must be a GP_Bag model type.
}
\examples{
### Toy Example in 2-dimensional case.
# Training set
set.seed(2)
n <- 40
x.train <- maximinLHS(n, 1)
y.train <- matrix(0, n, 1)
for(i in 1:n){ y.train[i] = (20*x.train[i])*sin(20*x.train[i]) + rnorm(length(x.train[i]),sd=1) }
dat.train <- data.frame(cbind(x.train,y.train))

# Test set
x.test <- seq(0, 1, 0.01);
y.test <- matrix(0, length(x.test), 1)
for(i in 1:length(x.test)){ y.test[i] = (20*x.test[i])*sin(20*x.test[i]) }
dat.test <- data.frame(cbind(x.test,y.test))

GP.mod <- GP_Bag(dat.train, x.test, y.test, iterations = 20, n = 40)

p <- plot_Bag(mod = GP.mod, x.test, y.test)
print(p)
}
\keyword{misc}

